User login implementation (#2)
* Add files via upload

Adding login user verification and creating functions to directly interact with the MySQL database of users

* Delete test.php

mistake to upload

* Complete website restructure

The entire website was restructured so that improvements could be made to allow for easy changes to website layout without modifying every page .HTML file. The function of user logins was also added with basic restrictions on where different users can access.

* Create folder

* Create folder

* Create folder

* Create folder

* Create folder

* Add files via upload

* Add files via upload

* Add files via upload

* Add files via upload

* Add files via upload

* Delete .temp

* Delete .temp

* Delete .temp

* Delete .temp

* Delete .temp

* Delete index.html

* Delete profile.html

* Delete login.html

* Delete search.html

* Delete stock.html

Desc.
The function of user's logging into the website has been implemented with the addition that they are restricted to pages that only they should be able to access. A side effect of trying to properly implement this was the complete restructuring of the website's architecture that also simplifies the websites URL to all the different web pages.