Creating user control of data (#6)
* Adding session start

* Adding session start

* Adding session start

* Adding session start

* Adding session start

* Adding session start

* Adding session start

* Adding session start

* Adding session start and controls

All desired controls have been implemented excepting for the modification of users or any file related operations.

* Adding session start

* Adding session start

Fixes some bugs highlighted when transitioning to a windows based development platform. The user's control of the data and structure of said data is implemented on a basic level. Technically, it is a usable product by all accounts. Finishing some desired functions, optimization and making it look better is the next step.