Add files via upload
Unit information functions are completed. Downloading the safe state is done, uploading a safe state is yet to be started. Mass stock functions also yet to be started. Home page was started but yet to be finalised. Export lists of users is also done so you can download a list of personnel who you have searched for.