Refactoring
Refactored most of the PHP and changed to HTML forms except in StockMC

As this was a long-term learning project, the refactoring of the code and the mostly the formatting of html and php code to better fit the standards needed to occur. A greater understanding of the possibilities and requirements often faced in web dev means a refactoring of the code would result greatly in a boost of maintainability for not just myself but possibly others.