Refactoring and Updates
Did a bit of refactoring, mostly just implemented the ability to edit the pls of each company (currently pre-defined) and the pl structure of each pl in terms of appointments. Also implemented the advanced search of years, coys and pls plus seeing the list of people who have been marked as graduated from the unit.